+++ Source/OnlineSubsystemAccelByte/Private/AsyncTasks/OnlineAsyncTaskAccelByteConnectLobby.cpp
@@ -114,16 +114,6 @@ void FOnlineAsyncTaskAccelByteConnectLobby::OnLobbyConnectSuccess()
 	}
 #endif
 
-	if (IdentityInterface.IsValid() && PartyInterface.IsValid())
-	{
-		AccelByte::Api::Lobby::FConnectionClosed OnLobbyConnectionClosedDelegate = AccelByte::Api::Lobby::FConnectionClosed::CreateStatic(FOnlineAsyncTaskAccelByteConnectLobby::OnLobbyConnectionClosed, LocalUserNum, IdentityInterface, PartyInterface);
-		ApiClient->Lobby.SetConnectionClosedDelegate(OnLobbyConnectionClosedDelegate);
-		
-		// #NOTE (Wiwing): Overwrite connect Lobby success delegate for reconnection
-		Api::Lobby::FConnectSuccess OnLobbyReconnectionDelegate = Api::Lobby::FConnectSuccess::CreateStatic(FOnlineAsyncTaskAccelByteConnectLobby::OnLobbyReconnected, LocalUserNum, IdentityInterface, PartyInterface);
-		ApiClient->Lobby.SetConnectSuccessDelegate(OnLobbyReconnectionDelegate);
-	}
-
 	CompleteTask(EAccelByteAsyncTaskCompleteState::Success);
 	AB_OSS_ASYNC_TASK_TRACE_END(TEXT("Sending request to get or create a default user profile for user '%s'!"), *UserId->ToDebugString());
 }
@@ -157,61 +147,6 @@ void FOnlineAsyncTaskAccelByteConnectLobby::OnLobbyDisconnectedNotif(const FAcce
 
 	UE_LOG_AB(Warning, TEXT("Lobby disconnected. Reason '%s'"), *Result.Message);
 	CompleteTask(EAccelByteAsyncTaskCompleteState::RequestFailed);
-}
-
-void FOnlineAsyncTaskAccelByteConnectLobby::OnLobbyConnectionClosed(int32 StatusCode, const FString& Reason, bool WasClean, int32 InLocalUserNum, const FOnlineIdentityAccelBytePtr IdentityInterface, const FOnlinePartySystemAccelBytePtr PartyInterface)
-{
-	UE_LOG_AB(Warning, TEXT("Lobby connection closed. Reason '%s' Code : '%d'"), *Reason, StatusCode);
-
-	if (!IdentityInterface.IsValid() || !PartyInterface.IsValid())
-	{
-		UE_LOG_AB(Warning, TEXT("Error due to either IdentityInterface or PartyInterface is invalid"));
-		return;
-	}
-
-	const TSharedPtr<const FUniqueNetId> UserIdPtr = IdentityInterface->GetUniquePlayerId(InLocalUserNum);
-	TSharedPtr<FUserOnlineAccount> UserAccount;
-
-	if (UserIdPtr.IsValid())
-	{
-		const FUniqueNetId& UserId = UserIdPtr.ToSharedRef().Get();
-		UserAccount = IdentityInterface->GetUserAccount(UserId);
-	}
-
-	if (UserAccount.IsValid())
-	{
-		const TSharedPtr<FUserOnlineAccountAccelByte> UserAccountAccelByte = StaticCastSharedPtr<FUserOnlineAccountAccelByte>(UserAccount);
-		UserAccountAccelByte->SetConnectedToLobby(false);
-	}
-
-	FString LogoutReason;
-	if (StatusCode == 1006) // Internet connection is disconnected
-	{
-		LogoutReason = TEXT("network-disconnection");
-		AccelByte::FApiClientPtr ApiClient = IdentityInterface->GetApiClient(InLocalUserNum);
-		ApiClient->CredentialsRef->ForgetAll();
-
-#if !AB_USE_V2_SESSIONS
-		TSharedPtr<FUniqueNetIdAccelByteUser const> LocalUserId = StaticCastSharedPtr<FUniqueNetIdAccelByteUser const>(IdentityInterface->GetUniquePlayerId(InLocalUserNum));
-		PartyInterface->RemovePartyFromInterface(LocalUserId.ToSharedRef());
-#endif
-	}
-	IdentityInterface->Logout(InLocalUserNum, LogoutReason);
-}
-
-void FOnlineAsyncTaskAccelByteConnectLobby::OnLobbyReconnected(int32 InLocalUserNum, const FOnlineIdentityAccelBytePtr IdentityInterface, const FOnlinePartySystemAccelBytePtr PartyInterface)
-{
-	UE_LOG_AB(Log, TEXT("Lobby successfully reconnected."));
-
-#if !AB_USE_V2_SESSIONS
-	if (IdentityInterface.IsValid() && PartyInterface.IsValid())
-	{
-		TSharedPtr<FUniqueNetIdAccelByteUser const> LocalUserId = StaticCastSharedPtr<FUniqueNetIdAccelByteUser const>(IdentityInterface->GetUniquePlayerId(InLocalUserNum));
-
-		PartyInterface->RemovePartyFromInterface(LocalUserId.ToSharedRef());
-		PartyInterface->RestoreParties(LocalUserId.ToSharedRef().Get(), FOnRestorePartiesComplete());
-	}
-#endif
 }
 
 void FOnlineAsyncTaskAccelByteConnectLobby::UnbindDelegates()