+++ Source/OnlineSubsystemAccelByte/Private/OnlineIdentityInterfaceAccelByte.cpp
@@ -15,7 +15,6 @@
 #include "Core/AccelByteError.h"
 #include "Core/AccelByteCredentials.h"
 #include "Api/AccelByteUserProfileApi.h"
-#include "Core/AccelByteOauth2Api.h"
 #include "OnlineSubsystemAccelByteModule.h"
 #include "GenericPlatform/GenericPlatformMisc.h"
 #include "Interfaces/OnlineExternalUIInterface.h"