+++ Source/OnlineSubsystemAccelByte/Public/OnlineSessionInterfaceV1AccelByte.h
@@ -25,6 +25,7 @@
 #include "FNboSerializeToBufferAccelByte.h"
 #include "OnlineDelegateMacros.h"
 #include "Runtime/Launch/Resources/Version.h"
+#include "AccelByteNetworkingStatus.h"
 
 class FOnlineSubsystemAccelByte;
 class FOnlineSessionV1AccelByte;
@@ -267,7 +268,7 @@ private:
 	FOnlineSessionSearchResult ConstructSessionResultForMatch(const FAccelBytePendingMatchInfo& PendingMatch);
 
 	void CreateP2PSession(FNamedOnlineSession* Session);
-	void OnRTCConnected(const FString& NetId, bool bWasSuccessful, FName SessionName);
+	void OnRTCConnected(const FString& NetId, const EAccelByteP2PConnectionStatus& Status, FName SessionName);
 	void OnRTCClosed(const FString & NetId, FName SessionName);
 	void OnSessionCreateSuccess(const FAccelByteModelsSessionBrowserData& Data, FName SessionName);
 	void UpdateSessionSettings(FName SessionName, FSessionSettings Settings);